Add shipping weight
We have to specify a unit of weight (lb, oz, g, kg) and although Spree
weights are unitless, active shipping defaults to lbs.

I don't want to introduce a dependency on active shipping, so I suggest
that we just make the easy assumption here and leave it to the user to
customize.