Add support to test image links with Active Storage
The current host is required for the disk service to work.

The test itself needed some modification because with Active Storage
the file needs to actually be loaded, while for Paperclip setting
the filename is enough to generate the URL. The test now really
upload the image on the Disk service.

We used an image from core for the upload, but blank.jpg has been
added in Solidus 2.11, while we have a thinking-cat.jpg in the api
module before. The conditional code addresses this.